rag_agent_exceldocs/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.1.0
langchain-huggingface>=0.0.2
langchain-openai>=0.0.1
pinecone>=2.2.1
pandas>=1.3.5

# Excel file support (various formats)
openpyxl>=3.0.10  # For modern Excel (.xlsx, .xlsm)
xlrd>=2.0.1  # For legacy Excel (.xls)
xlwt>=1.3.0  # For writing legacy Excel
odfpy>=1.4.1  # For OpenDocument (.ods)
pyxlsb>=1.0.10  # For Excel Binary (.xlsb)

# Embeddings and models
sentence-transformers>=2.2.2
openai>=1.1.1  # Used with custom base URL for aimlapi.com
torch>=2.0.0
transformers>=4.30.0

# Web app
streamlit>=1.27.0

# Document processing
python-magic>=0.4.27  # For file type detection
chardet>=5.1.0  # For character encoding detection

# Utilities
tqdm>=4.65.0  # For progress bars
python-dotenv>=1.0.0  # For environment variable management
pydantic>=1.10.8